Split ArrayFilterFunctionReturnTypeExtension to Helper
